Moved components to Modal folder
