+++ src/components/ourTeamSection/Modal/MemberCardModalBody.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image";
 import { useLocale, useTranslations } from "next-intl";
 import { LocaleType } from "@/types/LocaleType";
-import { SharePopover } from "../shared/SharePopover";
-import { NoImageHeart } from "../shared/NoImageHeart";
+import { SharePopover } from "../../shared/SharePopover";
+import { NoImageHeart } from "../../shared/NoImageHeart";
 import { MemberDataItemType } from "@/src/mockedData/membersData";
 import { MemberSocialLinks } from "./MemberSocialLinks";
 import { MemberProjectExperience } from "./MemberProjectExperience";
