+++ src/components/ourTeamSection/Modal/Modal.tsx
